Makepkgs - toolchain problem for ia32
Previously this was seen on Ubuntu (across lots of releases).  Now
I have access to the precursor Debian releases in my VM QA farm, I
see the problem is also present in Debian (not really a surprise).

If you build on/for ia32 with the default -O2, the resultant binaries
are bad ... the signature is that PCP QA for -g sanity fails for the
tests 087, 178, 184 and 338 ... turn off the compiler optimization, and
all these tests pass.  [there is a bigger class of failing tests
if you run the whole QA suite].

The problem is most accute for AMD CPUs, but I believe I've also seen
it on Intel CPUs ... this has been observed across 3 different machines.

This commit turns off gcc optimization for ia32 builds on Debian as
well as Ubuntu.