Error Injection Infrastracture for QA
As part of the coverage-driven changes to QA in PCP 4.0, it has become
apparent that we're going to need someway to exercise the "uncommon"
code paths associated with error detection and recovery.

These changes introduce an error injection infrastructure (for libpcp
only at this stage, alhough the mechanism is far more general and could
easily be extended).

If (and only if) libpcp is built with -DPM_FAULT_INJECTION then
the following become available via a new header "fault.h" that
is included in the source tree, but not installed

+ __pmFaultInject(char *ident, int class)
  Defines a fault point at which a fault of type class may be injected.
  - ident is a string to uniquely identify the fault point ... something
    like "libpcp/" __FILE__ ":<number>" works fine
  - class is a class of failure ... the following are currently
    defined:
    . PM_FAULT_ALLOC will cause the next call to malloc() or strcpy()
      (as proof of concept, other alloc routines could be added) to return
      NULL and set errno to ENOMEM.
    . PM_FAULT_PMAPI will cause the next call to a PMAPI routine (at the
      moment just __pmRegisterAnon() is supported as a proof of concept)
      to return PM_ERR_FAULT
  - Fault injection cannot be nested.  Each call to __pmFaultInject()
    clears any previous fault injection that has not yet been done.

+ PM_FAULT_POINT(ident, class)
  A macro, with PM_FAULT_INJECTION defined this becomes a call to
  __pmFaultInject(), otherwise PM_FAULT_POINT() emits no code.
  Throughout libpcp we use PM_FAULT_POINT() and NOT __pmFaultInject().

+ PM_FAULT_CHECK(class)
  A macro that emits no code unless PM_FAULT_INJECTION is defined, in
  which case if a fault of type class has been injected, the enclosing
  routine will return immediately with the value PM_ERR_FAULT ... assumes
  the enclosing routine is of type int foo(...) like all of the PMAPI
  routines.

+ __pmFaultSummary(void)
  Report a summary of fault points seen and faults injected.

Faults are ENABLED using a control file, identified by the environment
variable $PM_FAULT_CONTROL which contains lines of the form:
<ident> <op> <number>
that define injection guards.

This file (if it exists) is read the first time __pmFaultInject()
is called.  Within the control file, blank lines are ignored and lines
beginning with # are treated as comments.

<ident> is fault point string (as defined by a call to __pmFaultInject(),
or more usually the PM_FAULT_POINT() macro).
<op> is one of the C-style operators >=, >, ==, <, <=, !=, %
<number> is an unsigned integer
<op> <number> is optional, the default is >0

The semantics of the injection guards are: each time __pmFaultInject() is
called for a particular <ident>, a trip count is incremented (the first
trip is 1); if the C-style expression <tripcount> <op> <number> has the
value 1 (so true for most <op>s, remainder == 1 for the % <op>), then
a fault of the class defined for the fault point associated with <ident>
will be injected.

Additional tracing (via -Dfault and DBG_TRACE_FAULT) and a special
additional PMAPI error code (PM_ERR_FAULT) are also added, although
these will only ever be seen or used if libpcp has been built with
-DPM_FAULT_INJECTION.  If DBG_TRACE_FAULT is set the first time
__pmFaultInject() is called, then __pmFaultSummary() will be called
when the application exits (via atexit()).

Examples:
libpcp/src/derive.c:	PM_FAULT_CHECK()
libpcp/src/err.c:	PM_FAULT_POINT()
libpcp/src/events.c:	PM_FAULT_POINT()
libpcp/src/fault.c:	the underlying implementation
QA/477 and QA/478:	control file use