pmlogger - small problem with temporal index
As part of the cleanup logic at the end of writing a PCP archive,
pmlogger adds a final entry in the temporal index ... unfortunately
the entry's offset into the log volume was "not quite right"(tm).

The fix is trivial, and in practice I don't believe anything
wrong would have resulted from this.

Found while developing the new pmlogrewrite tool which _does_
write the final temporal index entry with correct offsets.