libpcp_pmda - add (long) opaque key support
Specifically pmdaCacheStoreKey() and pmdaCacheLookupKey() may be used to
provide "hints" to be used when assigning an internal instance identifier.
The "hints" are arbitrary sized binary data, so could much longer than
the 2^31-1 limit for an instance identifier.

Hashing ensures the hint -> instance identifier mapping is consistent
across uses and platforms with minimal chance of collisions, so this
provides a convenient way to map N-dimensional, 64-bit and other object
identifier schemes to instance identifiers within an instance domain.