environment games for openSUSE and systemctl/systemd
In openSUSE 12.1, /etc/rc.status intercepts our rc script and passes
control to systemctl which uses systemd ... the result is that messages
from our rc scripts are sent to syslog by default, and there is no
apparent way to revert to the classical behaviour.

This breaks much of the PCP QA.

This "hack" to rc-proc.sh allows PCP QA to set $PCPQA_NO_RC_STATUS and
continue to see stdout and stderr from our rc scripts.

To make all of this work, PCP QA needs to use "sudo -E" in place of "sudo"
and this leaves the one-trip guard in pcp.env set, but $PATH reset, so the
$PATH adjustment is now always done, and done in a better way, in pcp.env.