many - change rc start -> restart when appropriate
The rc script infrastructure in openSUSE 12.1 uses systemctl/systemd which
is stateful, so "start" is a NOP if the service is already "running".

But in many places, we've been using "start" as an alias for "restart"
(they are equivalent in all our scripts) so change to using "restart"
in all the places where that is the correct semantics.

Although we're avoiding systemctl/systemd in QA for openSUSE 12.1 for
other reasons, this change is to avoid being bitten at some point in
the future.