Demise of the binary format PMNS
Retire pmnscomp(1) and all references to the binary format PMNS.
Each call to pmLoadNameSpace(3) or pmLoadASCIINameSpace(3) will
load an ASCII format PMNS.

Added -N pmnsfile as an alternate to -n pmnsfile that allows duplicate
PMIDS ... currently only needed for pminfo and pmcd.

The binary format PMNS was one of a set of "premature" optimizations aimed
at minimizing the PCP overhead in the very early days of PCP evolution
... with pmcd now being the only application that really loads a PMNS,
and this happens only rarely, the binary PMNS adds complexity for no
real performance advantage, so it is being retired in PCP 4.0.