thread-safe pmErrStr & build libpcp with check-statics
In addition to creating pmErrStr_r(), and using this throughout
libpcp in lieu of pmErrStr(), this commit also introduces thread-safe
versons of the osstrerror_r() and netstrerror_r() macros to use and
hide the inconsistent GNU implementation of strerror_r().

Also included is the check-statics script as part of the build
to ensure additional static data items are not introduced without
appropriate thread-safe verification.