libpcp_gui/record.c - fix string not-terminated bug
In building rp->public->logfile, the optional addition of the filename
separator did not leave the result null-byte terminated, making the
following strcat() a loaded gun.

Found on openSUSE 12.1, although this has little to due with SUSE and
more to do with the state of the stack, the glibc version in place and
the alignment of the planets (signficantly, this bug was found and fixed
on the day of the Summer Solstice in the southern hemisphere).