pcp.spec.in - fileutils -> coreutils
At least on Mandriva 2011.0 there is no fileutils rpm, and the things
we depened on from fileutils (chgrp, chown, cp, ls, ...) all appear to
in coreutils, so change the Requires clause.

If this is not right for all the RPM platforms, then those that know
more than I will have to add vendor-specific variations.