More Mandriva man page hell
In 2011.0, the Mandriva pixies moved from .lzma compression for man
pages to .xz compression.  This changes the names of all the man page
files in the packages and the "dirt" in the build area ... wholesale
changes to configure.in and builddefs.in to accommodate the packaging
changes, then lots of .gitignore changes.