Assorted - botch in metric name lookup error handling
When pmNameLookup() is called with a list of metric names that
include a mix of derived metrics, dynamic metrics and unknown
metrics, the logic for handling the mixture of "odd" pmids
(PM_ID_NULL and pmid_domain(pmid) == DYNAMIC_PMID in particular)
was not always correct.

Found problems first in pmcd by accident during BSD pmda development,
then similar problems identified in dbpmda and libpcp.

QA/509 now checks for this type of botch.