Rearrange headers in the source tree
These changes allow all of PCP to be built in an environment where
/usr/include/pcp does not exist and there are no PCP libraries in
/usr/lib* ... so everything PCP-related that used in the build comes
out of the PCP source tree.

Further, the references to the PCP headers are consistent throughout.
If the source file is included in the binary package (by way of an
example or demo), then the form
	#include <pcp/pmapi.h>
is used to match the man pages.  Otherwise (for the vast majority of
files) the form
	#include "pmapi.h"
is used.

This style rule applies to all the PCP headers, not just pmapi.h.

With these changes, Mark's chroot build issue should be resolved.

Also in this commit, Perl build (Makefile.PL) changes to remove the
	Note (probably harmless): No library found for ...
messages when PCP libraries are not installed in /usr/lib* ... these
messages would not stop the build, but DID produce bad Perl modules
with unresolved symbols ... I've seen the messages in some of the QA
environments (but probably only for the first build!), and now
understand what they mean!!

Nathan, a similar Perl build fix needs to be made for the Mingw builds
that I cannot test at the moment.