pmcd - PMDA negotiation buffer pinning fix
Fallout from previous __pmUnpinPDUBuf() change in AgentNegotiate()
as part of the PCP 4.0 changes.  The logic here was wrong and we were
calling __pmUnpinPDUBuf() with a bogus buffer address on the end-of-file
error case (exposed by QA 427), which was finally caught by the addition
of the assert() in __pmUnpinPDUBuf() as part of commit 3534fbb.