misc - strange hangs associated with stderr redirection
First seen on LinuxMint 12 with PCP 4.0, where (apparently) random QA
tests were hanging on lines like
	some_cmd 2>&1 | some_filter
when some_filter was not seeing EOF when some_cmd exited!

Root cause never diagnosed, but replacing the line above with the
equivalent
	some_cmd >$tmp.tmp 2>&1
	some_filter <$tmp.tmp
worked a treat.

Sometimes it is best to just not ask "why?".