Fix an oversight in the pmda queue API regarding timestamps.
In the initial implementation of the shared PMDA queueing API
we've tried to maximise code sharing - a little too much so,
in this area.  We do not want to force the timestamp for an
event to have to be the time it arrived at the PMDA, the event
may encode its own (more accurate wrt the event) timestamp and
we should rather use that.

At this stage, its a trivial API tweak, and makes things alot
more flexible - we now pass the pmda-arrival-timestamp down to
the event decoder routine, allowing the PMDA to choose whether
or not it will use that or a timestamp inside the event buffer.

There is an associated pcpqa change to watch out for.  This API
(pmda queues) is new since last release, so there is no chance
that anyone else can be relying on this (other than pcpqa).