Update the postgresql PMDA to resolve issue with recovery metrics.
The initial version of this code was suboptimal wrt fetching the
postgresql recovery metrics on a database which is_in_recovery.  In
this mode, the current_xlog_position metric is unavailable, and the
way the SQL was structured, the fetch for all recovery metrics was
failing.  Split this into two separate value fetches and the values
are available always, whenever they exist (depends on host state in
terms of active/in_recovery).