Thread support - let the punter choose
Add --with-threads option to configure to allow punter decide if she wants
to have threads or not. It also serves as a helper for testing compilation
on platforms which may not have pthreads.

In the rest of the code enable thread support if we find a working pthread.h
and a mutex - the rest is of functionality is provided by libpcp.

And finally retire HAVE_THREAD_LOCALS in favour of HAVE___THREAD - the later
wins by majority.