Remove references to xpg4 id in start scripts.
This was confusing the startup script checkers, making
them believe there was an undeclared dependency on /usr
(which causes issues with mount/unmount order at system
start or stop, if not declared).

Removed, as such platform-specific path hackery belongs
in /etc/pcp.conf and/or /etc/pcp.env and we don't need
to support this old path anymore anyway.