Update the start script dependencies for /var use.
Warning message from lintian (this due to /var/tmp use):

"The given init script seems to refer to /var, possibly using a file from
there. Without a dependency on $local_fs in Required-Start or
Required-Stop, as appropriate, the init script might be run before /var
is mounted or after it's unmounted.

Using Should-Start or Should-Stop to declare the dependency is
conceptually incorrect since the $local_fs facility is always available.
Required-Start or Required-Stop should be used instead."

We were doing the latter, and we're now doing the former.