pmpost - refine environment handling
In an earlier life, pmpost was setuid root and we had an issue with
possible bogus environment variables.  The fix for that issue would not
allow $PCP_DIR nor $PCP_CONF to be passed in the environment, meaning
the relocation of the /var/log/pcp/NOTICES file via pcp.conf did not work.

pmpost is no longer setuid so allowing $PCP_CONF and $PCP_DIR to be
retained should be OK ... in the process fixed a bug where the special
case handling of $TZ was totally broken!