pmapi.h - unwind some of the recent unsigned changes
For event records, a couple of "array size" counts were changed
from int to unsigned int in commit acd24f294f20719f8838ad62042b33d246f8c15e
and this is not quite correct (these are not bit fields or flag fields).

The original design (although we may not have a use case yet in the code
base) was that these counters could be used to encode (negative) error
codes, similar to numpmid in a pmResult or numval in a pmValueSet.

But also, the expression p->ea_nrecords-1 became unsigned as a
by product of the change, which broke some of the code.

So revert ea_nrecords in pmEventArray and er_nparams in pmEventRecord
back to an int.