pmcd.conf - pathname silliness
As a result of the long and convoluted set of changes from early Irix
versions to the present day, we were not being consistent about how
pathnames for executables and DSOs are specified in pmcd.conf.

It all broke when $PCP_DIR was set in the aftermath of trying to run
a non-default installation after a configure --prefix=... build.

Now we're using absolute pathnames in the initial pmcd.conf with the
default PMDAs, in pmcd's parsing of pmcd.conf and in all the PMDA
Install procedures.