Makepkgs - better handling of $MAKE from environment
For FreeBSD, make is not gmake so we need the MAKE=... ./Makepkgs
recipe.  But the existing code in concert with configure.in was
inconsistent in this case and $PCP_MAKE_PROG became gmake rather than
the full path of gmake ... this ended up breaking the QA.

Replace ${MAKE:-make} here by something a little smarter.