libpcp - fix unaligned accesses for ia64
Noticed while debugging a different problem for QA 508 ... replace
copy-n-cast constructs in interp.c by correctly sized and aligned
temp var and memcpy() ... this only happens when diags are enabled.

And in util.c, just make the code a little clearer and similar to the
interp.c code (but no functional change here).