linux pmda - fix unaligned memory access for cgroup.subsys.hierarchy
Code was incorrectly using opaque pointer to private data with
the pmdaCache*() routines causing unaligned memory accesses on
ia64 platforms.

Although this problem is fixed, this code still seems a bit
dodgey and would appear to do the wrong thing if /proc/cgroups
ever contained more than one subsystem hierarchy with the same
subsystem name in more than one hierarchy (not sure if that
is possible).