pmlogger/pmlc/libpcp - fix segv in diag dark corner
pmlc and pmlogger use pmResults in a perverse ways to encode
logging state ... this was causing pmlogger to dump core when
it was running with full diagnostics enabled and pmlc enquired
about the logging state for event record metrics!

Found by accident when investigating a QA 139 failure on one host,
a new QA 510 now checks explicitly for this unusual combination.