pmlogger - bad lock recursion
This is a 3.6 regression as fallout from making libpcp thread-safe on
most platforms.  I checked all of the similar cases in libpcp, but missed
this one when an application (pmlogger) is using an internal libpcp API
that has locking side-effects.

We were calling __pmHandleToPtr() (which returns with the context locked)
and not being careful about releasing the context lock (c_lock) ... in the
worst case, this was in pmlogger's special fetch routine, so deeper
locking recursion every time we fetch metrics to be logged.

For many platforms, recursive locking seems to be allowed without obvious
limit (probably 2^32 or 2^64 depending on how the mutex is implemented),
but on Solaris (more specifically OpenIndiana) there appears to be a limit
of 500 or so ... which we can hit easily in QA.

The fix is to release the context lock as soon as we've finished poking
around with the context, and if necessary re-acquire the context pointer
and context lock later, then release it again ... repeat until bored.