Add lock trace debugging for libpcp.
This is _not_ enabled by default.  Need to #define PM_MULTI_THREAD_DEBUG
and rebuild libpcp.

When this is done, all PM_LOCK() and PM_UNLOCK() calls are
intercepted and optional diagnostics produced to report who is
locking/unlocking what and report unusual lock counts (except
for recursive locking we expect the lock count to be 0 before
PM_LOCK() and 1 before PM_UNLOCK()) ... reporting is enabled by
the [new] -Dlock debug option in concert with the -Dappl? options
to selectively enable diagnostics for particular lock classes -
appl0 for the libpcp global lock, appl1 for the per-context locks
and appl2 for everything else (including the ipc channel lock).

PCP QA is the only likely consumer of these features, unless we
have problems that _really_ look like they are lock-related.