pmlogextract - bad lock recursion
Similar to the pmlogger case, this is a 3.6 regression as fallout
from making libpcp thread-safe on most platforms.  I checked
all of the similar cases in libpcp, but missed this one when an
application (pmlogextract) is using an internal libpcp API that
has locking side-effects.

We were calling __pmHandleToPtr() (which returns with the context
locked) and not being careful about releasing the context lock
(c_lock) ... in the worst case, this was happening once per input
log record (data or metadata).

The fix is to release the context lock as soon as we've finished
poking around with the context.