pmlogextract - fix timestamp arithmetic
pmlogextract has been flawed from the outset in the choice of
doubles to hold all timestamps.  This involves integer to double
conversions from struct timeval and __pmTimeval timestamps which
cannot be accurate.

Eventually it blew up on OpenIndiana where timestamps that should
have been equal were slightly different and the code took a segv ...
this was killing QA as I could not even make the archives needed
for some of the QA tests.

Fixed by removing all the double timestamp arithmetic and operating
exclusively on tv_sec and tv_usec.