Further work on pmdabash
Redesign the mechanism for passing initial process-wide
state from bash to the PMDA - a single initial event to
transfer this (crucial) information is problematic.  It
is now stored in a second file alongside the traces, so
its always available (when PMDA restarts, for example).

Also simplified the way a new process trace is setup in
the PMDA, and added code to export the timestamps from
the trace.