Improvements to elasticsearch PMDA under load
Changes to pmdaelasticsearch to ensure it times out more
gracefully when elasticsearch doesn't respond in a timely
fashion.  This involves moving from LWP::Simple to now use
LWP::UserAgent, which allows connection timeouts to be set.

Before this change, a slow/hung response elasticsearch was
causing the PMDA to hang also, and then to be terminated by
pmcd (pipe closed).

Removed (out-of-date) check for the perl Switch module from
the Install script as well.