Rethink order of PATH setting for pcp start scripts
In Debian bug 677091 it is reported that the pcp start script
hangs the machine, owing to an unfortunate collision on the
name pmpost.  A binary with the same name also exists in the
"texlive-binaries" package, except it lives in /usr/bin, not
in $PCP_BINADM_DIR (hence no package install-time conflict).

This is most simply resolved by putting the pcp-specific PATH
components ahead of any system paths that might contain names
that conflict.