Add an interface supporting PMDAs doing dynamic queue allocation
The bash pmda creates a disposes of event queues on the fly, as
processes come and go.  There is a corner case where a client is
already talking to the PMDA, a new instrumented process starts,
and no events are delived.  This is because the new queue is not
aware of existing clients.

While there, cleanup the indom setup in libpcp_pmda open routine
to use globally available interface rather than open-coding some
things (around indom setup).  Also, have now added pmdabash into
default builds.