Rebuild (pmns) - changes for linux-proc PMDA split
Need to be selective about the top-level proc and cgroup metrics.
If we're on Linux and the proc PMDA is not there, need to cull these
ones ... this is needed for the migration/upgrade after the split
when the linux PMDA will be installed by default, but the proc PMDA
will not.
If the proc PMDA is installed in pmcd's config file, don't do the
culling.