configury: keep top-level configure in source repository and tarball
This is to allow platforms with only obsolete versions of autoconf
available to still use all the modern autoconf goodies -- as long as
the ./configure file was regenerated somewhere modern and saved.
Since autoconf-generated files aim to be backward-compatible (run on
ancient hosts), this is fine.  It also shortens the prerequisite list
and build time somewhat on build hosts.  The downside is keeping a
machine-generated file within git.

* configure: New file, incidentally generated on Fedora 16.
* .gitignore: Don't you dare ignore it!
* src/include/buildrules (*clean): Don't you dare delete it!
* GNUmakefile: Declare it a LSRCFILES= (and not CONFFILES=).
  (configure): Remove old target.
  (pcp.lsm etc.): Note to generate src/include/pcp/platform_defs.h
  also, to make sure ./configure run triggers before much else.
* Makepkgs, build/rpm/rpm.spec.in: Don't invoke autoconf any more.
* INSTALL, debian/control: Don't prereq it either.