configure.in: AC_PREREQ autoconf 2.60 or above
This is to preclude regeneration of ./configure using a too-old
autoconf that mistreats e.g. the AC_COMPILE_IFELSE construct.

I'd like to exclude rhel4's 2.59, but it's not clear at what point the
combination of AC_ARG_ENABLE / AC_COMPILE_IFELSE as used here has
started working properly, so the actually required version number
may be closer to the current 2.69.