Allow parent/child bash scripts to trace correctly
Resolve several remaining issues with a traced script starting
another.  In particular, the child retains the open trace fd,
but bash clears BASH_XTRACEFD from the environment.

Updated the sample script, and include it in sources, to demo
this functionality.  Also, resolved a build issue around use
of memalign - not supported on all platforms.