Add call backtrace for lock debug instrumentation
This is only used in QA, but needs to be in the mainline code base.
This change allows us to report the procedure call traceback at each
lock and unlock operation, which is most useful for lock recursion
and lock hierarchy inversion bugs should they arise.

None of this is compiled into the shipped versions of libpcp.