Add missing PDU length checks in __pmDecodeLogControl routine
__pmDecodeLogControl did not check the c_numpmid and v_numval fields
against the size of the PDU.  Due to the way the sizes passed to malloc
are calculated, heap objects could be allocated which are too small,
leading to a heap-based buffer overflow.

Original report and fixes reviewed by Florian Weimer of the Red Hat
Security team.  Red Hat bugzilla bug #841290.

Security advisory CVE-2012-3418.