Fix in-band signalling in __pmGetPDU resulting in memory leak
The return value of __pmGetPDU is both an error code and the value of
the type field of the PDU.  A negative type value is treated as an error
by HandleClientInput, and the PDU is never unpinned, even though it was
pinned by __pmGetPDU because there was no error.  This leads to a memory
leak and eventual pmcd crash.

Original report and fixes reviewed by Florian Weimer of the Red Hat
Security team.  Red Hat bugzilla bug #841319.

Security advisory CVE-2012-3420.