Fix a recent regression in log record decoding
The routines for decoding log records stored in PCP archives
share some code with the live host PDU decoding logic.  They
make use of __pmDecodeResult, in particular. So, before this
change (and with the result decoding fixes), they no longer
function because the buffer they pass in for decoding does
not have a proper PDU header.

Previously this was not a problem, but now the decode routine
looks at the PDU header to determine the size of the buffer
being decoded.  The fix is straight-forward - simply ensure
those header fields are initialised appropriately.