Ensure the __pmDecodeText ident parameter is always set
There was one obscure code path where the ident pointer would
not be set, if the type from the PDU was not valid, so ensure
its always initialised.  Situation handled in a similar way
to the way __pmEncodeText deals with an unexpected type field
(i.e. treat it as an indom).

Resolves Red Hat bugzilla bug #841306.