Tweaks to python Makefile to allow initial merging
This adds python sources into a LSRCFILES, which means
the Makepkgs build will be able to proceed (and they'll
appear in the source tarball now too).

The install target is renamed install_python (along the
lines of install_perl) so that this can be made optional
(and not default until packaging issues sorted out) as
will need to happen for platforms that don't have python
installed by default (MacOSX? Solaris? Win32 for sure).