Split the ChartItem into sampling and tracing components
Refactoring to allow for sampling charts (fixed size arrays
of values) to co-exist with tracing charts (which will have
variable sized sets of begin/end/point observations and not
fixed (sample) timestamps, but rather timestamps of when an
event occurred.

The base classes of Chart and ChartItem are now agnostic to
which style of chart they will be displaying.  Tons of work
to do still to get trace charts working well, but we're out
of the blocks.