Rework QWT5/6 differences in pmchart, to reach initial link
Started working through the many differences as we head to
Qwt6.  In particular; QwtPicker changes, some older types
removed (direct mapping to Qt classes), QwtPlotPrintFilter
is gone, and so on.

One largish area to sort out still is how missing data is
handled - we were using NaN values and a custom PlotCurve
draw routine, this is changed a fair bit and needs further
work.