Make python sources a first-class citizen in pcp
Several changes around python packaging; end result is
a new python-pcp RPM package and the first PCP client
tool in core pcp written in python (pmcollectl).

This builds on Franks GNUmakefile/setup.py fix to make
use of DIST_ROOT.  It works in a similar way to the
perl packages - build produces a file via the setup.py
--record option listing all the files in the package.
A little post-processing is required too, for rpm, as
it also adds byte-compiled python files (pyo), also.

Additionally, fix up the Group: tags for all of the
perl (and now python) packages to be more appropriate.

And finally, I moved pm-collectl.py into the top-level
src/pmcollect SUBDIR in pcp (renamed for consistency),
and install it to PCP_BIN_DIR as pmcollect - i.e. drop
the .py suffix, which seems to be "the done thing" by
most other RPM packages with python code in /usr/bin.