Fixes to get pcp-gui QA test 027 reliably passing
Two separate fixes were required here.  Firstly, we no longer
have a concept of PMDA_LOCAL_SAMPLE, so the first part of the
test is no longer sensible.  Update test to reflect.

However the second part of the test, where pmdumplog currently
dumps core, is very valid.  Turned out to be a bug in libpcp,
around handling of multiple contexts with a remote namespace.
This dumps core when trying to acquire a context-specific lock
due to a null pointer dereference on the c_pmcd->pc_lock field.
In local context mode, c_pmcd is null.  With the right (wrong?)
combination of local and remote contexts, a scenario where a
remote context lock is attempted on a local context.  Ugh!

In the process I tighten up the test for context equality in
one spot in the libqmc QmcGroup class.

All of the original pcp-gui tests now pass reliably for me.