Fix handling of multiple contexts with a remote namespace.
Found this issue when investgating a pmdumptext failure in 027
from the pcp-gui tests.

This dumps core when trying to acquire a context-specific lock
due to a null pointer dereference on the c_pmcd->pc_lock field.
In local context mode, c_pmcd is null.  With the right (wrong?)
combination of local and remote contexts, a scenario where a
remote context lock is attempted on a local context.

Root cause is an incorrect assignment during an if (...) test.
The braces were in the wrong position, causing the assignment
to not be the return value from __pmWhichContext, but rather
the result of the 'if (pmWhichContext() < 0)' test.  Argh!