Event tracking and reporting (debug) is now stable in libqmc
The commit completes the initial implementation.  We have the
ability to mimic pmevent reporting now via libqmc, and the QA
test program qmc_event (used in 029) does just that.  It also
exercises the capability of querying regular sampled metrics
as well as event trace metrics at the same time (i.e. within
a single fetch request).  Very funky!

The remaining issues were around things like modifying copies
of QVector iteration variables rather than underlying data,
that sort of thing.  Test 029 is updated to filter timestamps
which are now being reported (these change on each request to
pmdasample for new values, of course) as well as to include
the reported event parameters.  These are not filtered as we
use pmstore to setup for a deterministic stream of records.