Fix problem in sar2pcp handling of Fedora16 sar archives
This next problem is a genuine bug in the the special-case
code in the pmimport library dealing with instance names that
contain a space.

If instance names (say, virbr0 and virbr0-nic) are presented
that match up to either the space or the end of one (where one
is a substr of the other), then we will incorrectly believe
these to be the same instance name.  QA test 511 will be added
to once the failing case works (there are yet more problems in
there, onto the next...).

Red Hat Bugzilla: 859117