Fix problem in sar2pcp handling of down-rev (10->9) sar archives
This issue uncovered a bit of a design issue in our sar2pcp tool,
where it assumes that the output from the first sample is a good
representative sample for every other sample.  This falls down in
practice, because some tags are dynamic (network interfaces may
appear/disappear on the fly, as one observed example).

To resolve this, I've replaced the use of an indexed array of PMI
handles with a hash that maps metric names (+instance optionally)
to handles.  On first observation of any XML tag, the PMI handle
is created (not defined in hash), and all subsequent occurences
will use the cached handle.

Also in this commit is cleanup to use single-quotes for all non-
interpreted strings (most), and an option allowing XML files to
be passed into sar2pcp in place of binary files (heuristic used,
presence of .xml suffix seems a good indicator).  Mark requested
this to ease the pain of having to run sar2pcp on the same system
as sadf, which has to be the same system (same sysstat anyway) as
customers are using.

But wait, there's more...

Red Hat Bugzilla: 859117