Fix another problem in sar2pcp handling of ancient sar archives
Really old versions of sysstat used commas in place of decimal
points in the XML floating point numbers.  This causes failure
whenever perl attempts to convert that string to a float.  So,
we need to fix up the number (give it a decimal point instead
of a comma) before doing any arithmetic or attempting to store
the number.

This is the cause (if you should ever observe it) of this type
of perl warning/failure message:  (e.g.)
  'Argument "\x{31}\x{2c}..." isn't numeric in addition (+)'

Red Hat Bugzilla: 859117