Extend the import API to ease scripting language integration
Add two additional helper routines into libpcp_import to make
life easier for scripts.  Similar to the existing pmiUnits()
interface, we now also have pmiID() and pmiInDom() available
for high-level language wrappers.

Since I was extending the API, I took the opportunity to add
in the pmiErrStr_r() interface along the lines of the same in
core PMAPI, for obvious reasons (no future API changes should
be needed for thread-safe-ing if ever that is done).