Small extensions to sar2pcp command line handling
If SADF is set in the environment, use that over the default
(from PATH).  Allow upper/lower case .xml suffixes.  Improve
the handling of an empty file (or no data from sadf) - don't
pass the handle into XML::TokeParser because it spews a non-
intuitive error message in that case.