Pass updateValues() the actual update interval and visible points
The current pmchart internal updateValues call for each plot, on
each interval, had a wacky delta argument that was actually the
time scale visible interval and not the real (time) delta.  This
should be pushed down, and the scale calculated only in the one
spot that needs that, else confusion is sure to reign.