Work towards getting sensible values on a trace Y-axis
This is now working nicely, for certain definitions of nice.
When we get an event that indicates a parent, but no other
event identifying as said parent has arrived yet, some funky
behaviour results.  But otherwise, streams of events are now
identifiable on the left hand side.  Which is nice!