Updates to pmdabash, particularly improving script exit handling
Move the cleanup of the named pipe files into the PMDA, otherwise
we miss the final events.  Also take great care in handling the
actual unlink, as this results in instance domain updates (remove
completed process) and must be coordinated with attached clients
being sent the final events for that script.

Ensure the ID and PARENT flags remain set when marking END events.
Cull the pcp_trace instrumentation from the generated events, to
make the stream a bit clearer.