Refactor Chart class, splitting into Chart and ChartEngine
This is the second part of a large refactor to remove all
logic specific to either Sampling or Tracing into their
respective source files, such that the "core" Chart class
does not have to have a boolean flag saying which mode it
is currently operating under.

Previous large refactor in this area split out ChartItem
into separate tracing/sampling.  This now completes the
transition in the main Chart class (which covers all of
the ChartItems).