Workaround issues falling out of installing python scripts in QA
Firstly, ensure we install the python scripts into pcp-testsuite
package.  The obvious way to do that would install them with the
executable bit appropriately set.  However, that means RPM finds
them, and takes additional steps - like pre-compiling and adding
new files into the install root.  And the files installed depend
on both rpm and python versions.

So, leap over this rathole by using the same trick used by perl
scripts in QA - don't set the execute bits and run the scripts in
the tests with the interpreter prefix.  Seems the best of a bad
set of options.  Note that this is only done for the QA scripts
that are python code - the real python scripts in PCP (pmcollectl
for example) must be run without a prefix to properly test the
installed packages.